Создание примитивного интерфейса и рабочее приложение
